Utility/requirements.txt
dataclasses; python_version < '3.7'
lupa
chardet
colorama
pygame
Pillow
PyYAML